Refactor application structure and fix linting issues
- Extract HTTP server routing and file serving logic (`FileServer` struct and methods) into `server.go` to adhere to the Single Responsibility Principle.
- Move raw HTML template `HTML_PRELUDE` into `template.go` to keep static assets separate from business logic.
- Introduce `internal/reporter/reporter.go` for centralized error reporting and funnel HTTP internal errors through it.
- Fix multiple bugs in the POST handler, including nil pointer panics and defer issues.
- Resolve golangci-lint warnings and `govet` errors (e.g., using `%v` instead of `%w` in `log.Fatalf`).
